Move SystemConfig to its own cpp/h
