+++ source/ConfigFiles.h
@@ -15,80 +15,6 @@ static bool operator==(Array<T> a1, Array<T> a2) {
 	return !(a1 != a2);
 }
 
-/** System-specific configuration */
-class SystemConfig {
-public:
-	// Input parameters
-	bool	hasLogger = false;		///< Indicates that a hardware logger is present in the system
-	String	loggerComPort = "";		///< Indicates the COM port that the logger is on when hasLogger = True
-	bool	hasSync = false;		///< Indicates that a hardware sync will occur via serial card DTR signal
-	String	syncComPort = "";		///< Indicates the COM port that the sync is on when hasSync = True
-
-	SystemConfig() {};
-	/** Construct from Any */
-	SystemConfig(const Any& any) {
-		AnyTableReader reader(any);
-		int settingsVersion = 1;
-		reader.getIfPresent("settingsVersion", settingsVersion);
-		switch (settingsVersion) {
-		case 1:
-			reader.getIfPresent("hasLatencyLogger", hasLogger);
-			if (hasLogger) {
-				reader.get("loggerComPort", loggerComPort, "Logger COM port must be provided if \"hasLogger\" = true!");
-			}
-			else {
-				reader.getIfPresent("loggerComPort", loggerComPort);
-			}
-
-			reader.getIfPresent("hasLatencyLoggerSync", hasSync);
-			if (hasSync) {
-				reader.get("loggerSyncComPort", syncComPort, "Logger sync COM port must be provided if \"hasLoggerSync\" = true!");
-			}
-			else {
-				reader.getIfPresent("loggerSyncComPort", syncComPort);
-			}
-			break;
-		default:
-			debugPrintf("Settings version '%d' not recognized in SystemConfig.\n", settingsVersion);
-			break;
-		}	
-	}
-
-	static SystemConfig load(String filename = "systemconfig.Any") {
-		if (filename.empty()) { filename = "systemconfig.Any"; }
-		// Create default UserConfig file
-		if (!FileSystem::exists(System::findDataFile(filename, false))) { // if file not found, generate a default user config table
-			SystemConfig defConfig = SystemConfig();
-			defConfig.toAny().save(filename);						// Save the .any file
-			return defConfig;
-		}
-		return Any::fromFile(System::findDataFile(filename));
-	}
-
-	/** Serialize to Any */
-	Any toAny(const bool forceAll = true) const{
-		Any a(Any::TABLE);
-		SystemConfig def;
-		if(forceAll || def.hasLogger != hasLogger)			a["hasLatencyLogger"] = hasLogger;
-		if(forceAll || def.loggerComPort != loggerComPort)	a["loggerComPort"] = loggerComPort;
-		if(forceAll || def.hasSync != hasSync)				a["hasLatencyLoggerSync"] = hasSync;
-		if(forceAll || def.syncComPort != syncComPort)		a["loggerSyncComPort"] = syncComPort;
-		return a;
-	}
-
-	/** Print the latency logger config to log.txt */
-	void printToLog() {
-		const String loggerComStr = hasLogger ? loggerComPort : "None";
-		const String syncComStr = hasSync ? syncComPort : "None";
-		logPrintf("-------------------\nLDAT-R Config:\n-------------------\n\tLogger Present: %s\n\tLogger COM Port: %s\n\tSync Card Present: %s\n\tSync COM Port: %s\n\n",
-			hasLogger ? "True" : "False",
-			loggerComStr.c_str(),
-			hasSync ? "True" : "False",
-			syncComStr.c_str()
-		);
-	}
-};
-
 /**Class for managing user configuration*/
 class UserConfig {
 public:
