+++ vs/FPSci.lib.vcxproj.filters
@@ -69,6 +69,9 @@
     <ClInclude Include="..\source\SystemInfo.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\source\SystemConfig.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\source\sqlHelpers.cpp">
@@ -108,6 +111,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\source\SystemInfo.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\source\SystemConfig.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
